+++ apps/backend/medusa-config.ts
@@ -15,8 +15,8 @@ module.exports = defineConfig({
 			storeCors: process.env.STORE_CORS || 'http://localhost:5173',
 			adminCors: process.env.ADMIN_CORS || 'http://localhost:5173',
 			authCors: process.env.AUTH_CORS || 'http://localhost:5173',
-			jwtSecret: process.env.JWT_SECRET || 'supersecret',
-			cookieSecret: process.env.COOKIE_SECRET || 'supersecret',
+			jwtSecret: process.env.JWT_SECRET,
+			cookieSecret: process.env.COOKIE_SECRET,
 			restrictedFields: {
 				store: [
 					'complaint',