+++ packages/tax-lookup/src/providers/__tests__/provider.unit.spec.ts
@@ -1,3 +1,4 @@
+/// <reference types="jest" />
 import path from 'path'
 import { TaxLookupProvider, TaxLookupOptions } from '../../providers/tax-lookup/provider'